improved layout
