+++ src/components/c-vas-html-validation.vue
@@ -77,7 +77,7 @@
         border: 1px solid variables.$vas-color-status--error;
         background: color.adjust(variables.$vas-color-status--error, $lightness: 40%);
         color: variables.$vas-color-grayscale--0;
-        font-size: 0.8rem;
+        font-size: variables.$vas-font-size--small;
       }
     }
