+++ src/components/c-vas-navigation.vue
@@ -420,7 +420,7 @@
       position: relative;
       width: 100%;
       padding: variables.$vas-spacing--8 variables.$vas-spacing--30 variables.$vas-spacing--8 variables.$vas-spacing--8;
-      font-size: 14px;
+      font-size: variables.$vas-font-size--base;
     }
 
     &__search-clear-button {
