+++ src/components/c-vas-panel-action.vue
@@ -194,7 +194,7 @@
     padding: 0;
     gap: variables.$vas-spacing--4;
     font-family: inherit;
-    font-size: variables.$vas-font-size--12;
+    font-size: variables.$vas-font-size--small;
     line-height: 1;
     border-radius: variables.$vas-theme-border-radius;
     flex-shrink: 0;
