+++ src/elements/e-vas-checkbox.vue
@@ -177,7 +177,7 @@
     display: block;
     margin-bottom: 8px;
     cursor: pointer;
-    font-size: variables.$vas-font-size--16;
+    font-size: variables.$vas-font-size--label;
 
     &__field {
       position: absolute;
