+++ src/setup/scss/_basics.scss
@@ -12,46 +12,21 @@
   background-color: variables.$vas-color-grayscale--1000;
   color: variables.$vas-color-grayscale--0;
   font-family: variables.$vas-font-family--primary;
-  font-size: variables.$vas-font-size--16;
+  font-size: variables.$vas-font-size--base;
   line-height: 1.3;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 
-  h1 {
-    @include headline();
-
-    font-size: variables.$vas-font-size--30;
-  }
-
-  h2 {
-    @include headline();
-
-    font-size: variables.$vas-font-size--28;
-  }
-
-  h3 {
-    @include headline();
-
-    font-size: variables.$vas-font-size--26;
-  }
-
-  h4 {
-    @include headline();
-
-    font-size: variables.$vas-font-size--24;
-  }
-
-  h5 {
-    @include headline();
-
-    font-size: variables.$vas-font-size--20;
-  }
-
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
   h6 {
     @include headline();
 
-    font-size: variables.$vas-font-size--18;
-    }
+    font-size: variables.$vas-font-size--heading;
+  }
 
   a {
     color: variables.$vas-font-color--text;
