+++ src/setup/scss/_config.scss
@@ -1,4 +1,4 @@
 @use './variables';
 
-$config--font-size: 16px;
+$config--font-size: 12px;
 $config--line-height: 1.3;
