+++ src/styleguide/demo-pages/general/r-typography.vue
@@ -13,27 +13,37 @@
       </div>
       <div :class="b('right')">
         <div :class="b('section')">
-          <h1>h1.Headline</h1>
+          <div :class="b('scale-title')">Font Scale — Track A (Sidebar UI)</div>
+          <table :class="b('scale-table')">
+            <thead>
+              <tr>
+                <th>Token</th>
+                <th>Size</th>
+                <th>Preview</th>
+                <th>Used for</th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr
+                v-for="entry in fontScale"
+                :key="entry.token"
+              >
+                <td :class="b('scale-token')">{{ entry.token }}</td>
+                <td :class="b('scale-size')">{{ entry.size }}</td>
+                <td>
+                  <span :style="{ fontSize: entry.size }">{{ entry.sample }}</span>
+                </td>
+                <td :class="b('scale-usage')">{{ entry.usage }}</td>
+              </tr>
+            </tbody>
+          </table>
         </div>
+
         <div :class="b('section')">
-          <h2>h2.Headline</h2>
-        </div>
-        <div :class="b('section')">
-          <h3>h3.Headline</h3>
-        </div>
-        <div :class="b('section')">
-          <h4>h4.Headline</h4>
-        </div>
-        <div :class="b('section')">
-          <h5>h5.Headline</h5>
-        </div>
-        <div :class="b('section')">
-          <h6>h6.Headline</h6>
-        </div>
-        <div :class="b('section')">
-          <strong>&lt;p&gt; Paragraph</strong>
+          <h1>h1 — h6 Headline (all 15px)</h1>
           <p>{{ dummyText }}</p>
         </div>
+
         <div :class="b('section')">
           <p>&lt;a&gt; <a href="#">This is a link</a></p>
           <p>&lt;mark&gt; You can use the mark tag to <mark>highlight</mark> text.</p>
@@ -43,57 +53,7 @@
           <p>&lt;u&gt; <u>This line of text will render as underlined.</u></p>
           <p>&lt;small&gt; <small>This line of text is meant to be treated as fine print.</small></p>
           <p>&lt;strong&gt; <strong>This line rendered as bold text.</strong></p>
-          <p>&lt;b&gt; <b>This line rendered as bold text.</b></p>
           <p>&lt;em&gt; <em>This line rendered as italicized text.</em></p>
-          <p>&lt;abbr&gt; <abbr title="attribute">Abbreviations</abbr></p>
-        </div>
-
-        <strong>&lt;ul class="list-unordered"&gt; Unordered list</strong>
-        <div :class="b('section')">
-          <ul class="list-unordered">
-            <li
-              v-for="entry in listEntries"
-              :key="entry"
-            >
-              {{ line }}
-            </li>
-          </ul>
-        </div>
-
-        <strong>&lt;ol class="list-ordered"&gt; Unordered list</strong>
-        <div :class="b('section')">
-          <ol class="list-ordered">
-            <li
-              v-for="entry in listEntries"
-              :key="entry"
-            >
-              {{ line }}
-            </li>
-          </ol>
-        </div>
-
-        <strong>&lt;ul&gt; list no style</strong>
-        <div :class="b('section')">
-          <ul>
-            <li
-              v-for="entry in listEntries"
-              :key="entry"
-            >
-              {{ line }}
-            </li>
-          </ul>
-        </div>
-
-        <strong>&lt;ul class="list-inline"&gt; list inline</strong>
-        <div :class="b('section')">
-          <ul class="list-inline">
-            <li
-              v-for="entry in listEntries"
-              :key="entry"
-            >
-              {{ line }}
-            </li>
-          </ul>
         </div>
       </div>
     </div>
@@ -115,10 +75,13 @@
     data() {
       return {
         dummyText: faker.lorem.words(45),
-        listEntries: Array.from({ length: 5 })
-          .fill(1)
-          .map((entry) => Number(entry) + 1),
-        line: faker.lorem.words(20),
+        fontScale: [
+          { token: '$vas-font-size--tiny', size: '10px', sample: 'Badge label, icon annotation', usage: 'Badge labels, icon annotations, micro-captions' },
+          { token: '$vas-font-size--small', size: '11px', sample: 'Nav sub-item, code snippet, metadata', usage: 'Nav sub-items, code snippets, tag text, muted metadata' },
+          { token: '$vas-font-size--base', size: '12px', sample: 'Body text, form fields, nav items, tooltips', usage: 'Default body, form inputs, nav items, tooltips, markdown body' },
+          { token: '$vas-font-size--label', size: '13px', sample: 'Button text, form label, section label', usage: 'Button text, form labels, section labels, sidebar root base' },
+          { token: '$vas-font-size--heading', size: '15px', sample: 'Panel title, modal header', usage: 'All h1–h6 inside sidebar, panel/modal titles, page headers' },
+        ],
       };
     },
 
@@ -142,6 +105,8 @@
 </script>
 
 <style lang="scss">
+  @use '../../../setup/scss/variables';
+
   .r-typography {
     &__grid {
       display: grid;
@@ -169,14 +134,58 @@
     }
 
     &__font-special-chars {
-      font-size: 1rem;
+      font-size: variables.$vas-font-size--base;
       line-height: 1.4;
       overflow-wrap: break-word;
     }
 
     &__section {
       margin-bottom: 2rem;
+      padding: variables.$vas-spacing--16;
       background-color: #fbfbfb;
+    }
+
+    &__scale-title {
+      font-size: variables.$vas-font-size--label;
+      font-weight: variables.$vas-font-weight--bold;
+      margin-bottom: variables.$vas-spacing--12;
+    }
+
+    &__scale-table {
+      width: 100%;
+      border-collapse: collapse;
+      font-size: variables.$vas-font-size--base;
+
+      th,
+      td {
+        text-align: left;
+        padding: variables.$vas-spacing--6 variables.$vas-spacing--12;
+        border-bottom: 1px solid var(--vas-theme-border-color);
+        vertical-align: middle;
+      }
+
+      th {
+        font-size: variables.$vas-font-size--tiny;
+        font-weight: variables.$vas-font-weight--bold;
+        text-transform: uppercase;
+        letter-spacing: 0.08em;
+        color: var(--vas-theme-text-color-muted);
+      }
+    }
+
+    &__scale-token {
+      font-family: monospace;
+      font-size: variables.$vas-font-size--small;
+      white-space: nowrap;
+    }
+
+    &__scale-size {
+      font-weight: variables.$vas-font-weight--bold;
+      white-space: nowrap;
+    }
+
+    &__scale-usage {
+      color: var(--vas-theme-text-color-muted);
     }
   }
 </style>
